Set compound filetypes separately, then all together
This tricks any `FileType` autocommands to run for each component, and
then the compound filetype works to run both syntax files.

More information: <https://vi.stackexchange.com/questions/4893/>

The downside is that it means that you can't use something like

    autocmd FileType pandoc ...

to mean that an autocommand should **only** run for `pandoc` filetypes,
not any compound filetype that it's a part of. You'd have to add another
autocommand like

    autocmd FileType pandoc.*,*.pandoc.*,*.pandoc ...

to **undo** any setup that came from the bare `pandoc` autocommands.
I think it's much more common to want `FileType pandoc` to fire
universally, so I'm fine with inflicting this more verbose setup on
anyone wanting to do the more rare thing.